Reorganized schema files.
