+++ db/schemas/operations/__init__.py
@@ -1,0 +1,4 @@
+from db.schemas.operations.create import *  # noqa
+from db.schemas.operations.drop import *  # noqa
+from db.schemas.operations.select import *  # noqa
+from db.schemas.operations.update import *  # noqa
