+++ db/schemas/operations/select.py
@@ -1,0 +1,47 @@
+import warnings
+
+from sqlalchemy import MetaData, select, and_, not_, or_, Table
+
+from db import types
+
+
+TYPES_SCHEMA = types.base.SCHEMA
+EXCLUDED_SCHEMATA = [TYPES_SCHEMA, "information_schema"]
+
+
+def reflect_schema(engine, name=None, oid=None):
+    # If we have both arguments, the behavior is undefined.
+    try:
+        assert name is None or oid is None
+    except AssertionError as e:
+        raise e
+    metadata = MetaData()
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore", message="Did not recognize type")
+        pg_namespace = Table("pg_namespace", metadata, autoload_with=engine)
+    sel = (
+        select(pg_namespace.c.oid, pg_namespace.c.nspname.label("name"))
+        .where(or_(pg_namespace.c.nspname == name, pg_namespace.c.oid == oid))
+    )
+    with engine.begin() as conn:
+        schema_info = conn.execute(sel).fetchone()
+    return schema_info
+
+
+def get_mathesar_schemas_with_oids(engine):
+    metadata = MetaData()
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore", message="Did not recognize type")
+        pg_namespace = Table("pg_namespace", metadata, autoload_with=engine)
+    sel = (
+        select(pg_namespace.c.nspname.label('schema'), pg_namespace.c.oid)
+        .where(
+            and_(
+                *[pg_namespace.c.nspname != schema for schema in EXCLUDED_SCHEMATA],
+                not_(pg_namespace.c.nspname.like("pg_%"))
+            )
+        )
+    )
+    with engine.begin() as conn:
+        result = conn.execute(sel).fetchall()
+    return result
