+++ db/tables/operations/create.py
@@ -2,7 +2,8 @@ from sqlalchemy import Column, String, Table, MetaData
 from sqlalchemy.ext import compiler
 from sqlalchemy.schema import DDLElement
 
-from db import columns, schemas
+from db import columns
+from db.schemas import operations as schema_operations
 
 
 def create_mathesar_table(name, schema, columns_, engine, metadata=None):
@@ -12,7 +13,7 @@ def create_mathesar_table(name, schema, columns_, engine, metadata=None):
     table.
     """
     columns_ = columns.init_mathesar_table_column_list_with_defaults(columns_)
-    schemas.create_schema(schema, engine)
+    schema_operations.create_schema(schema, engine)
     # We need this so that we can create multiple mathesar tables in the
     # same MetaData, enabling them to reference each other in the
     # SQLAlchemy context (e.g., for creating a ForeignKey relationship)
