+++ mathesar/tests/imports/test_csv.py
@@ -7,7 +7,8 @@ from sqlalchemy import text
 from mathesar.models import DataFile, Schema
 from mathesar.errors import InvalidTableError
 from mathesar.imports.csv import create_table_from_csv, get_sv_dialect, get_sv_reader
-from db.schemas import create_schema, get_schema_oid_from_name
+from db.schemas.operations.create import create_schema
+from db.schemas.utils import get_schema_oid_from_name
 
 TEST_SCHEMA = "import_csv_schema"
